Remove dedicated /jobs route, use homepage with type=job filter
- Delete /jobs route files (+page.svelte, +page.server.ts, jobs.remote.ts)
- Keep /jobs/submit for job submission flow
- Move applyToJob function to job/[slug]/job.remote.ts
- Update all /jobs links to /?type=job:
  - SidebarJobs "View all" link
  - Job detail page "Back to all jobs" link
  - Checkout success/cancel pages

Jobs are now viewed on the homepage with ?type=job filter,
consistent with other content types like recipes, videos, etc.
